refactor(noti): use inject provider
