+++ packages/core/src/constant.ts
@@ -1,4 +1,8 @@
-import type { NotiOptions, NotiPositionTuple } from './types'
+import type { InjectionKey } from 'vue'
+
+import type { NotiContext, NotiOptions, NotiPositionTuple } from './types'
+
+export const INJECT_KEY: InjectionKey<NotiContext> = Symbol('noti')
 
 export const DEFAULT_SETTING: NotiOptions = {
   message: '',
